+++ test/test_helper.rb
@@ -1,3 +1,4 @@
+require 'bundler/setup'
 require 'minitest/autorun'
 require 'active_support/testing/declarative'
 require 'minitest-activemodel'