+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.0.2 - unreleased
 
++ Add support for Rubinius.
 + Add `validate_uniqueness_of` matcher.
 + Add `scoped_to` option to `validate_uniqueness_of` matcher.
 + Add `case_insensitive` option to `validate_uniqueness_of` matcher.